Add EnumEventData derive
