+++ event-sauce-derive/src/derives/event_data.rs
@@ -1,70 +1,12 @@
-use proc_macro2::Span;
+use super::{parse_event_data_attributes, EventDataAttributes};
 use quote::quote;
-use syn::{Attribute, Data, DataStruct, DeriveInput, Fields, FieldsNamed, Meta, NestedMeta, Path};
-
-/// Attempt to assign a value to a variable, failing if the variable is already populated.
-///
-/// Prevents attributes from being defined twice
-macro_rules! try_set {
-    ($i:ident, $v:expr, $t:expr) => {
-        match $i {
-            None => $i = Some($v),
-            Some(_) => return Err(syn::Error::new_spanned($t, "duplicate attribute")),
-        }
-    };
-}
-
-macro_rules! fail {
-    ($t:expr, $m:expr) => {
-        return Err(syn::Error::new_spanned($t, $m));
-    };
-}
+use syn::{Data, DataStruct, DeriveInput, Fields, FieldsNamed};
 
 enum BuilderType {
     Create,
     Update,
     Delete,
     Purge,
-    Action,
-}
-
-struct EventDataAttributes {
-    entity: Path,
-}
-
-fn parse_event_data_attributes(input: &[Attribute]) -> syn::Result<EventDataAttributes> {
-    let mut entity = None;
-
-    for attr in input {
-        let meta = attr
-            .parse_meta()
-            .map_err(|e| syn::Error::new_spanned(attr, e))?;
-
-        match meta {
-            Meta::List(list) if list.path.is_ident("event_sauce") => {
-                for value in list.nested.iter() {
-                    match value {
-                        NestedMeta::Meta(meta) => match meta {
-                            Meta::Path(path) => try_set!(entity, path.clone(), path),
-
-                            u => fail!(u, "unexpected attribute"),
-                        },
-                        u => fail!(u, "unexpected attribute"),
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    let entity = entity.ok_or_else(|| {
-        syn::Error::new(
-            Span::call_site(),
-            "Attribute entity is required, e.g. #[event_sauce(entity = User)",
-        )
-    })?;
-
-    Ok(EventDataAttributes { entity })
 }
 
 fn expand_derive_event_data_struct(
@@ -95,10 +37,6 @@ fn expand_derive_event_data_struct(
             quote!(event_sauce::PurgeEntityBuilder),
             quote!(event_sauce::PurgeEventBuilder),
         ),
-        BuilderType::Action => (
-            quote!(event_sauce::ActionEntityBuilder),
-            quote!(event_sauce::ActionEventBuilder),
-        ),
     };
 
     Ok(quote!(
@@ -114,27 +52,6 @@ fn expand_derive_event_data_struct(
 
         impl #builder_impl<#ident> for #entity {}
     ))
-}
-
-fn expand_derive_event_data_enum(
-    input: &DeriveInput,
-    builder_type: BuilderType,
-) -> syn::Result<proc_macro2::TokenStream> {
-    let ident = &input.ident;
-
-    let EventDataAttributes { entity } = parse_event_data_attributes(&input.attrs)?;
-
-    if matches!(builder_type, BuilderType::Action) {
-        let builder_impl = quote!(event_sauce::ActionEventBuilder);
-        Ok(quote!(
-            impl #builder_impl<#ident> for #entity {}
-        ))
-    } else {
-        Err(syn::Error::new_spanned(
-            input,
-            "enums shall use action-builder only",
-        ))
-    }
 }
 
 pub fn expand_derive_create_event_data(
@@ -201,30 +118,6 @@ pub fn expand_derive_delete_event_data(
         }) => expand_derive_event_data_struct(input, BuilderType::Delete),
 
         Data::Enum(_) => Err(syn::Error::new_spanned(input, "enums are not supported")),
-
-        Data::Union(_) => Err(syn::Error::new_spanned(input, "unions are not supported")),
-    }
-}
-
-pub fn expand_derive_action_event_data(
-    input: &DeriveInput,
-) -> syn::Result<proc_macro2::TokenStream> {
-    match &input.data {
-        Data::Struct(DataStruct {
-            fields: Fields::Named(FieldsNamed { .. }),
-            ..
-        })
-        | Data::Struct(DataStruct {
-            fields: Fields::Unnamed(_),
-            ..
-        })
-        | Data::Struct(DataStruct {
-            fields: Fields::Unit,
-            ..
-        }) => expand_derive_event_data_struct(input, BuilderType::Action),
-
-        // TODO: this was added by me
-        Data::Enum(_) => expand_derive_event_data_enum(input, BuilderType::Action),
 
         Data::Union(_) => Err(syn::Error::new_spanned(input, "unions are not supported")),
     }
